requirements.txt
numpy
tqdm
opencv-python
Pillow

torch
torchvision

git+https://github.com/openai/CLIP.git